refactor column alignment code
repo_name_width should actually be the longest combination of
"repo/pkgname" for all upgrades, not the longest repo plus longest
pkgname in isolation. This can be tested against using a really long
repo name with a short pkgname, and vice-versa.

Also refactor the logic to avoid various off-by-one corrections and make
the code easier to understand.